Renamed extension to expansion
