+++ core/src/main/java/me/wolfyscript/utilities/expansions/PackMetaFile.java
@@ -16,7 +16,7 @@
  *     along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package me.wolfyscript.utilities.extensions;
+package me.wolfyscript.utilities.expansions;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
